Refactor: cleaned up files
