+++ packages/markers/marker.go
@@ -2,11 +2,13 @@ package markers
 
 import (
 	"fmt"
+	"sort"
 	"strconv"
 	"strings"
 	"sync"
 
 	"github.com/iotaledger/hive.go/cerrors"
+	"github.com/iotaledger/hive.go/datastructure/thresholdmap"
 	"github.com/iotaledger/hive.go/marshalutil"
 	"github.com/iotaledger/hive.go/stringify"
 	"golang.org/x/xerrors"
@@ -384,6 +386,319 @@ func (m *Markers) SequenceToString() (s string) {
 
 // endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
 
+// region ReferencingMarkers ///////////////////////////////////////////////////////////////////////////////////////////
+
+// ReferencingMarkers is a data structure that allows to denote which Markers of child Sequences in the Sequence DAG
+// reference a given Marker in a Sequence.
+type ReferencingMarkers struct {
+	referencingIndexesBySequence MarkerReferences
+	mutex                        sync.RWMutex
+}
+
+// NewReferencingMarkers is the constructor for the ReferencingMarkers.
+func NewReferencingMarkers() (referencingMarkers *ReferencingMarkers) {
+	referencingMarkers = &ReferencingMarkers{
+		referencingIndexesBySequence: make(map[SequenceID]*thresholdmap.ThresholdMap),
+	}
+
+	return
+}
+
+// ReferencingMarkersFromBytes unmarshals ReferencingMarkers from a sequence of bytes.
+func ReferencingMarkersFromBytes(referencingMarkersBytes []byte) (referencingMarkers *ReferencingMarkers, consumedBytes int, err error) {
+	marshalUtil := marshalutil.New(referencingMarkersBytes)
+	if referencingMarkers, err = ReferencingMarkersFromMarshalUtil(marshalUtil); err != nil {
+		err = xerrors.Errorf("failed to parse ReferencingMarkers from MarshalUtil: %w", err)
+		return
+	}
+	consumedBytes = marshalUtil.ReadOffset()
+
+	return
+}
+
+// ReferencingMarkersFromMarshalUtil unmarshals ReferencingMarkers using a MarshalUtil (for easier unmarshaling).
+func ReferencingMarkersFromMarshalUtil(marshalUtil *marshalutil.MarshalUtil) (referencingMarkers *ReferencingMarkers, err error) {
+	referencingMarkers = &ReferencingMarkers{
+		referencingIndexesBySequence: make(map[SequenceID]*thresholdmap.ThresholdMap),
+	}
+
+	referencingMarkers.referencingIndexesBySequence, err = MarkerReferencesFromMarshalUtil(marshalUtil, thresholdmap.UpperThresholdMode)
+	return referencingMarkers, err
+}
+
+// Add adds a new referencing Marker to the ReferencingMarkers.
+func (r *ReferencingMarkers) Add(index Index, referencingMarker *Marker) {
+	r.mutex.Lock()
+	defer r.mutex.Unlock()
+
+	thresholdMap, thresholdMapExists := r.referencingIndexesBySequence[referencingMarker.SequenceID()]
+	if !thresholdMapExists {
+		thresholdMap = thresholdmap.New(thresholdmap.UpperThresholdMode)
+		r.referencingIndexesBySequence[referencingMarker.SequenceID()] = thresholdMap
+	}
+
+	thresholdMap.Set(uint64(index), referencingMarker.Index())
+}
+
+// Get returns the Markers of child Sequences that reference the given Index.
+func (r *ReferencingMarkers) Get(index Index) (referencingMarkers *Markers) {
+	r.mutex.RLock()
+	defer r.mutex.RUnlock()
+
+	referencingMarkers = NewMarkers()
+	for sequenceID, thresholdMap := range r.referencingIndexesBySequence {
+		if referencingIndex, exists := thresholdMap.Get(uint64(index)); exists {
+			referencingMarkers.Set(sequenceID, referencingIndex.(Index))
+		}
+	}
+
+	return
+}
+
+// Bytes returns a marshaled version of the ReferencingMarkers.
+func (r *ReferencingMarkers) Bytes() (marshaledReferencingMarkers []byte) {
+	r.mutex.RLock()
+	defer r.mutex.RUnlock()
+
+	marshalUtil := marshalutil.New()
+	marshalUtil.WriteUint64(uint64(len(r.referencingIndexesBySequence)))
+	for sequenceID, thresholdMap := range r.referencingIndexesBySequence {
+		marshalUtil.Write(sequenceID)
+		marshalUtil.WriteUint64(uint64(thresholdMap.Size()))
+		thresholdMap.ForEach(func(node *thresholdmap.Element) bool {
+			marshalUtil.WriteUint64(node.Key().(uint64))
+			marshalUtil.WriteUint64(uint64(node.Value().(Index)))
+
+			return true
+		})
+	}
+
+	return marshalUtil.Bytes()
+}
+
+// String returns a human readable version of the ReferencingMarkers.
+func (r *ReferencingMarkers) String() (humanReadableReferencingMarkers string) {
+	r.mutex.RLock()
+	defer r.mutex.RUnlock()
+
+	indexes := make([]Index, 0)
+	referencingMarkersByReferencingIndex := make(map[Index]*Markers)
+	for sequenceID, thresholdMap := range r.referencingIndexesBySequence {
+		thresholdMap.ForEach(func(node *thresholdmap.Element) bool {
+			index := Index(node.Key().(uint64))
+			referencingIndex := node.Value().(Index)
+			if _, exists := referencingMarkersByReferencingIndex[index]; !exists {
+				referencingMarkersByReferencingIndex[index] = NewMarkers()
+
+				indexes = append(indexes, index)
+			}
+
+			referencingMarkersByReferencingIndex[index].Set(sequenceID, referencingIndex)
+
+			return true
+		})
+	}
+	sort.Slice(indexes, func(i, j int) bool {
+		return indexes[i] < indexes[j]
+	})
+
+	for i, index := range indexes {
+		for j := i + 1; j < len(indexes); j++ {
+			referencingMarkersByReferencingIndex[indexes[j]].ForEach(func(referencingSequenceID SequenceID, referencingIndex Index) bool {
+				if _, exists := referencingMarkersByReferencingIndex[index].Get(referencingSequenceID); exists {
+					return true
+				}
+
+				referencingMarkersByReferencingIndex[index].Set(referencingSequenceID, referencingIndex)
+
+				return true
+			})
+		}
+	}
+
+	thresholdStart := "0"
+	referencingMarkers := stringify.StructBuilder("ReferencingMarkers")
+	for _, index := range indexes {
+		thresholdEnd := strconv.FormatUint(uint64(index), 10)
+
+		if thresholdStart == thresholdEnd {
+			referencingMarkers.AddField(stringify.StructField("Index("+thresholdStart+")", referencingMarkersByReferencingIndex[index]))
+		} else {
+			referencingMarkers.AddField(stringify.StructField("Index("+thresholdStart+" ... "+thresholdEnd+")", referencingMarkersByReferencingIndex[index]))
+		}
+
+		thresholdStart = strconv.FormatUint(uint64(index)+1, 10)
+	}
+
+	return referencingMarkers.String()
+}
+
+// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+// region ReferencedMarkers ////////////////////////////////////////////////////////////////////////////////////////////
+
+// ReferencedMarkers is a data structure that allows to denote which Marker of a Sequence references which other Markers
+// of its parent Sequences in the Sequence DAG.
+type ReferencedMarkers struct {
+	referencedIndexesBySequence MarkerReferences
+	mutex                       sync.RWMutex
+}
+
+// NewReferencedMarkers is the constructor for the ReferencedMarkers.
+func NewReferencedMarkers(markers *Markers) (referencedMarkers *ReferencedMarkers) {
+	referencedMarkers = &ReferencedMarkers{
+		referencedIndexesBySequence: make(map[SequenceID]*thresholdmap.ThresholdMap),
+	}
+
+	initialSequenceIndex := markers.HighestIndex() + 1
+	markers.ForEach(func(sequenceID SequenceID, index Index) bool {
+		thresholdMap := thresholdmap.New(thresholdmap.LowerThresholdMode)
+		thresholdMap.Set(uint64(initialSequenceIndex), index)
+
+		referencedMarkers.referencedIndexesBySequence[sequenceID] = thresholdMap
+
+		return true
+	})
+
+	return
+}
+
+// ReferencedMarkersFromBytes unmarshals ReferencedMarkers from a sequence of bytes.
+func ReferencedMarkersFromBytes(parentReferencesBytes []byte) (referencedMarkers *ReferencedMarkers, consumedBytes int, err error) {
+	marshalUtil := marshalutil.New(parentReferencesBytes)
+	if referencedMarkers, err = ReferencedMarkersFromMarshalUtil(marshalUtil); err != nil {
+		err = xerrors.Errorf("failed to parse ReferencedMarkers from MarshalUtil: %w", err)
+		return
+	}
+	consumedBytes = marshalUtil.ReadOffset()
+
+	return
+}
+
+// ReferencedMarkersFromMarshalUtil unmarshals ReferencedMarkers using a MarshalUtil (for easier unmarshaling).
+func ReferencedMarkersFromMarshalUtil(marshalUtil *marshalutil.MarshalUtil) (referencedMarkers *ReferencedMarkers, err error) {
+	referencedMarkers = &ReferencedMarkers{
+		referencedIndexesBySequence: make(map[SequenceID]*thresholdmap.ThresholdMap),
+	}
+
+	referencedMarkers.referencedIndexesBySequence, err = MarkerReferencesFromMarshalUtil(marshalUtil, thresholdmap.LowerThresholdMode)
+	return referencedMarkers, err
+}
+
+// Add adds new referenced Markers to the ReferencedMarkers.
+func (r *ReferencedMarkers) Add(index Index, referencedMarkers *Markers) {
+	r.mutex.Lock()
+	defer r.mutex.Unlock()
+
+	referencedMarkers.ForEach(func(referencedSequenceID SequenceID, referencedIndex Index) bool {
+		thresholdMap, exists := r.referencedIndexesBySequence[referencedSequenceID]
+		if !exists {
+			thresholdMap = thresholdmap.New(thresholdmap.LowerThresholdMode)
+			r.referencedIndexesBySequence[referencedSequenceID] = thresholdMap
+		}
+
+		thresholdMap.Set(uint64(index), referencedIndex)
+
+		return true
+	})
+}
+
+// Get returns the Markers of parent Sequences that were referenced by the given Index.
+func (r *ReferencedMarkers) Get(index Index) (referencedMarkers *Markers) {
+	r.mutex.RLock()
+	defer r.mutex.RUnlock()
+
+	referencedMarkers = NewMarkers()
+	for sequenceID, thresholdMap := range r.referencedIndexesBySequence {
+		if referencedIndex, exists := thresholdMap.Get(uint64(index)); exists {
+			referencedMarkers.Set(sequenceID, referencedIndex.(Index))
+		}
+	}
+
+	return
+}
+
+// Bytes returns a marshaled version of the ReferencedMarkers.
+func (r *ReferencedMarkers) Bytes() (marshaledReferencedMarkers []byte) {
+	r.mutex.RLock()
+	defer r.mutex.RUnlock()
+
+	marshalUtil := marshalutil.New()
+	marshalUtil.WriteUint64(uint64(len(r.referencedIndexesBySequence)))
+	for sequenceID, thresholdMap := range r.referencedIndexesBySequence {
+		marshalUtil.Write(sequenceID)
+		marshalUtil.WriteUint64(uint64(thresholdMap.Size()))
+		thresholdMap.ForEach(func(node *thresholdmap.Element) bool {
+			marshalUtil.WriteUint64(node.Key().(uint64))
+			marshalUtil.WriteUint64(uint64(node.Value().(Index)))
+
+			return true
+		})
+	}
+
+	return marshalUtil.Bytes()
+}
+
+// String returns a human readable version of the ReferencedMarkers.
+func (r *ReferencedMarkers) String() (humanReadableReferencedMarkers string) {
+	r.mutex.RLock()
+	defer r.mutex.RUnlock()
+
+	indexes := make([]Index, 0)
+	referencedMarkersByReferencingIndex := make(map[Index]*Markers)
+	for sequenceID, thresholdMap := range r.referencedIndexesBySequence {
+		thresholdMap.ForEach(func(node *thresholdmap.Element) bool {
+			index := Index(node.Key().(uint64))
+			referencedIndex := Index(node.Value().(uint64))
+			if _, exists := referencedMarkersByReferencingIndex[index]; !exists {
+				referencedMarkersByReferencingIndex[index] = NewMarkers()
+
+				indexes = append(indexes, index)
+			}
+
+			referencedMarkersByReferencingIndex[index].Set(sequenceID, referencedIndex)
+
+			return true
+		})
+	}
+	sort.Slice(indexes, func(i, j int) bool {
+		return indexes[i] < indexes[j]
+	})
+
+	for i, referencedIndex := range indexes {
+		for j := 0; j < i; j++ {
+			referencedMarkersByReferencingIndex[indexes[j]].ForEach(func(sequenceID SequenceID, index Index) bool {
+				if _, exists := referencedMarkersByReferencingIndex[referencedIndex].Get(sequenceID); exists {
+					return true
+				}
+
+				referencedMarkersByReferencingIndex[referencedIndex].Set(sequenceID, index)
+
+				return true
+			})
+		}
+	}
+
+	referencedMarkers := stringify.StructBuilder("ReferencedMarkers")
+	for i, index := range indexes {
+		thresholdStart := strconv.FormatUint(uint64(index), 10)
+		thresholdEnd := "INF"
+		if len(indexes) > i+1 {
+			thresholdEnd = strconv.FormatUint(uint64(indexes[i+1])-1, 10)
+		}
+
+		if thresholdStart == thresholdEnd {
+			referencedMarkers.AddField(stringify.StructField("Index("+thresholdStart+")", referencedMarkersByReferencingIndex[index]))
+		} else {
+			referencedMarkers.AddField(stringify.StructField("Index("+thresholdStart+" ... "+thresholdEnd+")", referencedMarkersByReferencingIndex[index]))
+		}
+	}
+
+	return referencedMarkers.String()
+}
+
+// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
+
 // region markersByRank ////////////////////////////////////////////////////////////////////////////////////////////////
 
 // markersByRank is a collection of Markers that groups them by the rank of their Sequence.
@@ -574,6 +889,75 @@ func (m *markersByRank) String() (humanReadableMarkersByRank string) {
 	}
 
 	return structBuilder.String()
+}
+
+// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+// region markerReferences /////////////////////////////////////////////////////////////////////////////////////////////
+
+type MarkerReferences map[SequenceID]*thresholdmap.ThresholdMap
+
+// MarkerReferencesFromMarshalUtil unmarshals MarkerReferences using a MarshalUtil (for easier unmarshaling).
+func MarkerReferencesFromMarshalUtil(marshalUtil *marshalutil.MarshalUtil, mode thresholdmap.Mode) (referenceMarkers MarkerReferences, err error) {
+	referenceMarkers = make(map[SequenceID]*thresholdmap.ThresholdMap)
+
+	sequenceCount, err := marshalUtil.ReadUint64()
+	if err != nil {
+		err = xerrors.Errorf("failed to parse Sequence count (%v): %w", err, cerrors.ErrParseBytesFailed)
+		return
+	}
+	for i := uint64(0); i < sequenceCount; i++ {
+		sequenceID, sequenceIDErr := SequenceIDFromMarshalUtil(marshalUtil)
+		if sequenceIDErr != nil {
+			err = xerrors.Errorf("failed to parse SequenceID from MarshalUtil: %w", sequenceIDErr)
+			return
+		}
+
+		referenceCount, referenceCountErr := marshalUtil.ReadUint64()
+		if referenceCountErr != nil {
+			err = xerrors.Errorf("failed to parse reference count (%v): %w", referenceCountErr, cerrors.ErrParseBytesFailed)
+			return
+		}
+		thresholdMap := thresholdmap.New(mode)
+		switch mode {
+		case thresholdmap.LowerThresholdMode:
+			for j := uint64(0); j < referenceCount; j++ {
+				referencingIndex, referencingIndexErr := marshalUtil.ReadUint64()
+				if referencingIndexErr != nil {
+					err = xerrors.Errorf("failed to read referencing Index (%v): %w", referencingIndexErr, cerrors.ErrParseBytesFailed)
+					return
+				}
+
+				referencedIndex, referencedIndexErr := marshalUtil.ReadUint64()
+				if referencedIndexErr != nil {
+					err = xerrors.Errorf("failed to read referenced Index (%v): %w", referencedIndexErr, cerrors.ErrParseBytesFailed)
+					return
+				}
+
+				thresholdMap.Set(referencingIndex, Index(referencedIndex))
+			}
+		case thresholdmap.UpperThresholdMode:
+			for j := uint64(0); j < referenceCount; j++ {
+				referencedIndex, referencedIndexErr := marshalUtil.ReadUint64()
+				if referencedIndexErr != nil {
+					err = xerrors.Errorf("failed to read referenced Index (%v): %w", referencedIndexErr, cerrors.ErrParseBytesFailed)
+					return
+				}
+
+				referencingIndex, referencingIndexErr := marshalUtil.ReadUint64()
+				if referencingIndexErr != nil {
+					err = xerrors.Errorf("failed to read referencing Index (%v): %w", referencingIndexErr, cerrors.ErrParseBytesFailed)
+					return
+				}
+
+				thresholdMap.Set(referencedIndex, Index(referencingIndex))
+			}
+		}
+
+		referenceMarkers[sequenceID] = thresholdMap
+	}
+
+	return referenceMarkers, err
 }
 
 // endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
